requirements.txt
# venv Source
# sqllite3
# kubernetes
# pymysql
mysql.connector
# subprocess